+++ backend/reconciliation.py
@@ -1,70 +1,13 @@
 # external libraries
-from urllib.parse import parse_qs, quote
-import re
-from flask import request, Response
-import requests
 from SPARQLWrapper import SPARQLWrapper, POST, JSON, DIGEST
 from rdflib import Graph, URIRef, Literal
 from rdflib.namespace import SDO, RDFS, OWL
-import os
 
-MYLINKSET = 'http://localhost:9999/bigdata/sparql'
 UPDATEMYLINKSET = 'http://localhost:9999/blazegraph/namespace/kb/sparql/update'
 LILNKSETGRAPH = 'http://w3id.org/polifonia/linkset/'
 LINKSET_FILE = 'linkset.nt'
 
 WHITE_LIST = ['wikidata', 'dbpedia', 'viaf', 'discogs']
-
-
-def __run_query_string(active, query_string,
-                       is_post=False, content_type="application/x-www-form-urlencoded"):
-    try:
-        query_str_decoded = query_string.decode('utf-8')
-    except Exception as e:
-        query_str_decoded = query_string
-    parsed_query = parse_qs(query_str_decoded)
-
-    if query_str_decoded is None or query_str_decoded.strip() == "":
-        print('solve 1')
-        # return render_template('sparql.html',active=active)
-
-    if re.search("updates?", query_str_decoded, re.IGNORECASE) is None:
-        if "query" in parsed_query or "select" in query_str_decoded.lower():
-            return __contact_tp(query_string, is_post, content_type)
-        else:
-            print('solve 2')
-            # return render_template('sparql.html',active=active)
-    else:
-        print('solve 3')
-        # return render_template('403.html'), 403
-
-
-def __contact_tp(data, is_post, content_type):
-    accept = request.args.get('HTTP_ACCEPT')
-    if accept is None or accept == "*/*" or accept == "":
-        accept = "application/sparql-results+json"
-
-    data = data if isinstance(data, bytes) else quote(data)
-    if is_post:
-        req = requests.post(MYLINKSET, data=data,
-                            headers={'content-type': content_type, "accept": accept})
-    else:
-        req = requests.get("%s?query=%s" % (MYLINKSET, data),
-                           headers={'content-type': content_type, "accept": accept})
-
-    if req.status_code == 200:
-        response = Response()
-        response.headers['Access-Control-Allow-Origin'] = '*'
-        response.headers['Access-Control-Allow-Credentials'] = 'true'
-        response.headers['Content-Type'] = req.headers["content-type"]
-        response.mimetype = "application/sparql-results+json"
-        return req.json()
-    else:
-        print('solve 4')
-        # return render_template('error.html',
-        #                        status_code=str(req.status_code),
-        #                        headers={"Content-Type": request.content_type},
-        #                        text=req.text)
 
 
 # altro parametro tipo lista con sparqlenpoint
